chore(log-export): remove the superseded sync download endpoint
