+++ backend/ee/onyx/configs/license_enforcement_config.py
@@ -78,7 +78,7 @@ PATH_PREFIX_MIN_TIER: dict[str, Tier] = {
     "/manage/admin/standard-answer": Tier.ENTERPRISE,
     "/admin/token-rate-limits": Tier.ENTERPRISE,
     "/admin/hooks": Tier.ENTERPRISE,  # outbound webhooks
-    "/admin/log-export": Tier.ENTERPRISE,  # container-local log download
+    "/admin/log-export": Tier.ENTERPRISE,  # deployment-wide log export
     "/analytics": Tier.ENTERPRISE,  # non-admin analytics (e.g. assistant stats)
     "/evals": Tier.ENTERPRISE,
     "/scim": Tier.ENTERPRISE,  # SCIM protocol
